+++ olsync/olbrowserlogin.py
@@ -6,12 +6,13 @@
 # Description: Overleaf Browser Login Utility
 # Author: Moritz Glöckl
 # License: MIT
-# Version: 1.1.6
+# Version: 1.1.5
 ##################################################
 
-from PyQt5.QtCore import *
-from PyQt5.QtWidgets import *
-from PyQt5.QtWebEngineWidgets import *
+from PySide6 import QtCore
+from PySide6 import QtWidgets
+from PySide6 import QtWebEngineWidgets
+from PySide6 import QtWebEngineCore
 
 # Where to get the CSRF Token and where to send the login request to
 LOGIN_URL = "https://www.overleaf.com/login"
@@ -22,7 +23,7 @@ JAVASCRIPT_CSRF_EXTRACTOR = "document.getElementsByName('ol-csrfToken')[0].conte
 COOKIE_NAMES = ["overleaf_session2", "GCLB"]
 
 
-class OlBrowserLoginWindow(QMainWindow):
+class OlBrowserLoginWindow(QtWidgets.QMainWindow):
     """
     Overleaf Browser Login Utility
     Opens a browser window to securely login the user and returns relevant login data.
@@ -31,22 +32,22 @@ class OlBrowserLoginWindow(QMainWindow):
     def __init__(self, *args, **kwargs):
         super(OlBrowserLoginWindow, self).__init__(*args, **kwargs)
 
-        self.webview = QWebEngineView()
+        self.webview = QtWebEngineWidgets.QWebEngineView()
 
         self._cookies = {}
         self._csrf = ""
         self._login_success = False
 
-        self.profile = QWebEngineProfile(self.webview)
+        self.profile = QtWebEngineCore.QWebEngineProfile(self.webview)
         self.cookie_store = self.profile.cookieStore()
         self.cookie_store.cookieAdded.connect(self.handle_cookie_added)
-        self.profile.setPersistentCookiesPolicy(QWebEngineProfile.NoPersistentCookies)
+        self.profile.setPersistentCookiesPolicy(QtWebEngineCore.QWebEngineProfile.NoPersistentCookies)
 
-        self.profile.settings().setAttribute(QWebEngineSettings.JavascriptEnabled, True)
+        self.profile.settings().setAttribute(QtWebEngineCore.QWebEngineSettings.JavascriptEnabled, True)
 
-        webpage = QWebEnginePage(self.profile, self)
+        webpage = QtWebEngineCore.QWebEnginePage(self.profile, self)
         self.webview.setPage(webpage)
-        self.webview.load(QUrl.fromUserInput(LOGIN_URL))
+        self.webview.load(QtCore.QUrl.fromUserInput(LOGIN_URL))
         self.webview.loadFinished.connect(self.handle_load_finished)
 
         self.setCentralWidget(self.webview)
@@ -56,11 +57,12 @@ class OlBrowserLoginWindow(QMainWindow):
         def callback(result):
             self._csrf = result
             self._login_success = True
-            QCoreApplication.quit()
+            print('success')
+            QtCore.QCoreApplication.quit()
 
         if self.webview.url().toString() == PROJECT_URL:
             self.webview.page().runJavaScript(
-                JAVASCRIPT_CSRF_EXTRACTOR, callback
+                JAVASCRIPT_CSRF_EXTRACTOR, 0,  callback
             )
 
     def handle_cookie_added(self, cookie):
@@ -82,7 +84,7 @@ class OlBrowserLoginWindow(QMainWindow):
 
 
 def login():
-    app = QApplication([])
+    app = QtWidgets.QApplication([])
     ol_browser_login_window = OlBrowserLoginWindow()
     ol_browser_login_window.show()
     app.exec()
@@ -91,3 +93,5 @@ def login():
         return None
 
     return {"cookie": ol_browser_login_window.cookies, "csrf": ol_browser_login_window.csrf}
+
+
